[refactor] 구독한 언론사 없을 때 페이지 그리는 함수 분리
