+++ asset/data/constants.js
@@ -7,5 +7,5 @@ const FILTER_TYPE = {
     ALL: "all",
     SUBSCRIBED: "subscribed",
 }
-const GRID_ITEMS_COUNT = 24;
-export {CATEGORY_LIST, VIEW_TYPE, FILTER_TYPE, GRID_ITEMS_COUNT}
+const GRID_ITEMS_PER_PAGE = 24;
+export {CATEGORY_LIST, VIEW_TYPE, FILTER_TYPE, GRID_ITEMS_PER_PAGE}
