+++ script/grid-view/grid.js
@@ -1,16 +1,13 @@
 import { store } from "../../store/store.js";
-import { FILTER_TYPE } from "../../asset/data/constants.js";
+import { FILTER_TYPE, GRID_ITEMS_PER_PAGE } from "../../asset/data/constants.js";
+import { drawEmptySubView } from "../view-utils/empty-sub-view.js";
 
 const gridContainer = document.querySelector(".grid-box");
 const pressCover = document.querySelector(".press-cover");
-const emptyView = document.querySelector(".empty-view");
 const subBtn = document.querySelector(".sub-btn");
 const unsubBtn = document.querySelector(".unsub-btn");
 
-function drawEmptyGrid() {
-    gridContainer.classList.add("hide");
-    emptyView.classList.remove("hide");
-}
+
 function getGridData() {
     let gridData;
     let {crntFilter} = store.getViewState();
@@ -50,10 +47,10 @@ function drawGrid(){
     let {crntPage} = store.getViewState()
     let imgIdxList = getGridData();
     if (imgIdxList.length === 0) {
-        drawEmptyGrid();
+        drawEmptySubView();
     }
     gridContainer.innerHTML = "";
-    for (let i=24*crntPage;i<24*(crntPage+1);i++){
+    for (let i=GRID_ITEMS_PER_PAGE*crntPage;i<GRID_ITEMS_PER_PAGE*(crntPage+1);i++){
         gridContainer.innerHTML += `
             <li class="pressItem" index=${imgIdxList[i]}>
             ${i >= imgIdxList.length ? "" : `<img src="./asset/logo/light/img${imgIdxList[i]}.svg" />` }
